+++ test/jmap/email/email_object_test.dart
@@ -12,6 +12,7 @@ import 'package:jmap_dart_client/jmap/mail/email/email_address.dart';
 import 'package:jmap_dart_client/jmap/mail/email/email_body_part.dart';
 import 'package:jmap_dart_client/jmap/mail/email/email_body_value.dart';
 import 'package:jmap_dart_client/jmap/mail/email/email_header.dart';
+import 'package:jmap_dart_client/jmap/mail/email/email_header_value.dart';
 import 'package:jmap_dart_client/jmap/mail/email/individual_header_identifier.dart';
 import 'package:jmap_dart_client/jmap/mail/email/keyword_identifier.dart';
 import 'package:jmap_dart_client/jmap/mail/mailbox/mailbox.dart';
@@ -358,7 +359,7 @@ void main() {
         headers: {
           EmailHeader('Return-Path', ' <dphamhoang@linagora.com>')
         },
-        headerUserAgent: {IndividualHeaderIdentifier.headerUserAgent: 'Team-Mail/0.7.8 Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/110.0.0.0 Safari/537.36'}
+        individualHeaders: {IndividualHeaderIdentifier.headerUserAgent: const TextHeaderValue('Team-Mail/0.7.8 Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/110.0.0.0 Safari/537.36')}
       );
 
       final parsedEmail = Email.fromJson(jsonDecode(emailObjectAsString));